Add files via upload
Disclaimer for Issue Asset Function:

Disclaimer: Creating and whitelisting your token is not synonymous
with endorsement by the Krosscoin Project and its partner
companies. You also agree to not use your token as securities but
only for rewards, payments of goods and services and utility
purposes. Consult a legal team if it is intended to be used otherwise.
Running this function signifies agreement to disclaimer above